% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clccommunities.R
\name{add_community_data}
\alias{add_community_data}
\title{Get a table of community data about a vector of postal codes}
\usage{
add_community_data(dat, postal_code_colname = NULL)
}
\arguments{
\item{postal_code}{}
}
\value{

}
\description{
Get a table of community data about a vector of postal codes
}
