% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clccommunities.R
\name{clc_calculate_low_income}
\alias{clc_calculate_low_income}
\title{For a vector of postal codes, identify
those which are in low income communities}
\usage{
clc_calculate_low_income(
  postal_code,
  lim_at_prevalence_threshold = 0.193,
  format = c("label", "indicator")
)
}
\arguments{
\item{postal_code}{A vector of postal codes}

\item{lim_at_prevalence_threshold}{The lowest prevalence of LIM-AT that would categorize a community as low income,
Defaults to 19.3\%, the lower bound of the 4th quartile of LIM-AT prevalence among FSAs.}

\item{format}{The format of the return vector. Either "label" (the default), or "indicator".
If the format is "label", the returned vector will be a factor with verbose category labels.
If the format is "indicator", the returned vector will be a numeric boolean (1, 0, or NA)}
}
\value{
A vector of low income assessments for each value in postal_code, in the specified format
}
\description{
For a vector of postal codes, identify
those which are in low income communities
}
