% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clccommunities.R
\name{clc_calculate_rural}
\alias{clc_calculate_rural}
\title{For a vector of postal codes, identify which ones are in
rural communities}
\usage{
clc_calculate_rural(postal_code, sac_index_threshold = 4, format = "label")
}
\arguments{
\item{postal_code}{A vector of postal codes}

\item{sac_index_threshold}{The lowest SAC index that should be classified as rural. Defaults to 4 (moderate metropolitan influence zone)}

\item{format}{The format of the return vector. Either "label" (the default), or "indicator".
If the format is "label", the returned vector will be a factor with verbose category labels.
If the format is "indicator", the returned vector will be a numeric boolean (1, 0, or NA)}
}
\value{

}
\description{
For a vector of postal codes, identify which ones are in
rural communities
}
